Share with NativeAOT and Mono
