+++ src/coreclr/nativeaot/System.Private.CoreLib/src/System.Private.CoreLib.csproj
@@ -172,7 +172,6 @@
     <Compile Include="System\Helpers.cs" />
     <Compile Include="System\Math.NativeAot.cs" />
     <Compile Include="System\MathF.NativeAot.cs" />
-    <Compile Include="System\MissingMemberException.NativeAot.cs" />
     <Compile Include="System\MulticastDelegate.cs" />
     <Compile Include="System\Object.NativeAot.cs" />
     <Compile Include="System\Reflection\MissingMetadataException.cs" />
