--- src/mono/System.Private.CoreLib/src/System/MissingMemberException.Mono.cs
@@ -1,15 +1,0 @@
-// Licensed to the .NET Foundation under one or more agreements.
-// The .NET Foundation licenses this file to you under the MIT license.
-
-namespace System
-{
-    public partial class MissingMemberException
-    {
-#pragma warning disable IDE0060
-        internal static string FormatSignature(byte[] signature)
-        {
-            return "";
-        }
-#pragma warning restore IDE0060
-    }
-}
